requirements.txt
python-dotenv==1.0.1
langchain==0.3.3
langchain-community==0.3.2
langchain-openai==0.2.2
pydantic==2.9.2
faiss-cpu==1.8.0.post1
chromadb==0.5.5
pypdf==5.0.1
tiktoken==0.7.0
fastapi==0.115.0
uvicorn==0.30.6
streamlit==1.38.0
